+++ doc/source/releasenotes.rst
@@ -4,6 +4,9 @@
 Release notes
 =============
 
+1.0.4
+======
+
 1.0.3
 ======
 * Getting thermodynamic quantities in the SGC MC now also retrieves averages from observers.